requirements.txt
# torch
the_utils>=0.8.0
graph_datasets>=0.14.1
torch_geometric==2.6.1
